# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "컬컴 백오피스 REST API 문서",
        "title": "Culcom Backoffice API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@culcom.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/calendar/create-event": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "구글 캘린더에 새로운 이벤트를 생성합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "integrations"
                ],
                "summary": "구글 캘린더 이벤트 생성",
                "parameters": [
                    {
                        "description": "캘린더 이벤트 정보",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/integrations.CreateCalendarEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/calendar/disconnect": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "구글 캘린더의 연동을 해제하고 토큰을 삭제합니다",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "integrations"
                ],
                "summary": "구글 캘린더 연동 해제",
                "responses": {
                    "200": {
                        "description": "구글 캘린더 연동이 해제되었습니다",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/external/sms": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "SMS 테스트 메시지를 발송합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "integrations"
                ],
                "summary": "SMS 테스트 발송",
                "parameters": [
                    {
                        "description": "SMS 발송 정보",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/integrations/activate": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "특정 서비스의 연동을 활성화합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "integrations"
                ],
                "summary": "연동 활성화",
                "parameters": [
                    {
                        "description": "service_id",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "활성화되었습니다",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/integrations/disconnect": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "특정 서비스의 연동을 해제(비활성화)합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "integrations"
                ],
                "summary": "연동 해제",
                "parameters": [
                    {
                        "description": "service_id",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "연결이 해제되었습니다",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/service/sms": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "고객에게 SMS 메시지를 전송하고 이력을 저장합니다",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "SMS 메시지 전송",
                "parameters": [
                    {
                        "type": "string",
                        "description": "고객 시퀀스",
                        "name": "customer_seq",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "발신번호",
                        "name": "sender_phone",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "수신번호",
                        "name": "receiver_phone",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "메시지 내용",
                        "name": "message",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "인증 실패",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "integrations.CreateCalendarEventRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "description": "코멘트",
                    "type": "string"
                },
                "customer_name": {
                    "description": "고객 이름",
                    "type": "string"
                },
                "duration": {
                    "description": "소요시간 (분, 기본값 60분)",
                    "type": "integer"
                },
                "interview_date": {
                    "description": "인터뷰 일시 (YYYY-MM-DD HH:MM:SS)",
                    "type": "string"
                },
                "phone_number": {
                    "description": "전화번호",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "SessionAuth": {
            "type": "apiKey",
            "name": "user-session",
            "in": "cookie"
        }
    }
}